test: Extract tests for BuildStatusGeneratorMixin
